backend/requirements.dev.txt
SQLAlchemy
typing-extensions
fastapi
pydantic
starlette
email-validator
pydantic-settings
python-multipart
uvicorn[standard]
websockets
sqlmodel
pyjwt
cryptography
passlib
psycopg
psycopg2-binary
requests
orjson
typish
dnspython
ecdsa
email_validator
fastapi-mail
Jinja2
MarkupSafe
phonenumbers
pluggy
pyasn1
pydantic
pydantic-extra-types
pydantic-settings
pydantic_core
pytest
pytest-asyncio
python-dotenv
python-engineio
python-jose
python-multipart
python-socketio
requests
simple-websocket
six
sniffio
stripe
typing_extensions
urllib3
uvicorn
alembic
bcrypt==4.0.1
google-generativeai